+++ src/lib/ember-cli-onedata-common/addon/services/session.js
@@ -227,10 +227,10 @@ export default SessionCore.extend({
       this.get('onWebSocketError'),
       this.get('onWebSocketClose'),
       isPublic
-    ).then(({ oneproviderHostname, oneproviderToken }) =>
+    ).then(({ oneproviderApiOrigin, oneproviderToken }) =>
       safeExec(this, 'setProperties', {
         oneproviderToken,
-        oneproviderHostname,
+        oneproviderApiOrigin,
       })
     );